refactor(PoweredByIcons): move powered-by-icons.tsx file to data folder & update import path
