+++ apps/client/src/data/powered-by-icons.tsx
